Moved geometry refs into its own file
