+++ src/ref.js
@@ -1,0 +1,50 @@
+export const Geometry = {
+  SHORT_HOUSE: {
+    path: 'window_house', // for test
+    obj: {}
+  },
+  TOWER: {
+    path: 'tall_tower', // for test
+    obj: {},
+  },
+  FLOOR_APT: {
+    path: 'floor',
+    obj: {},
+    sizeRatio: 1.7,
+  },
+  GROUND_FLOOR_APT: {
+    path: 'floor',
+    obj: {},
+  },
+  ROOF_APT: {
+    path: 'apartment_roof', 
+    obj: {}
+  },
+  FLOOR_SKY: {
+    path: 'skyscraper_floor',
+    obj: {},
+    sizeRatio: 4,
+  },
+  GROUND_FLOOR_SKY: {
+    path: 'skyscraper_floor',
+    obj: {},
+  },
+  ROOF_SKY: {
+    path: 'skyscraper_roof', //TODO: replace
+    obj: {}
+  },
+  PARK: {
+    path: 'park',
+    obj: {},
+  },
+  TREE: {
+    path: 'tree',
+    obj:{},
+  }
+  // STORE_FRONT: { // -> can function as GROUND_FLOOR_APT or GROUND_FLOOR_SKY
+  //   path: '', //TODO
+  //   obj: {},
+  // },
+
+  //TODO: STORE_FRONT
+}
